Fixed some bugs

+++ Bai thuc hanh so 1/bai1.cpp
@@ -11,7 +11,7 @@ int main(){
   int tong=a+b;
   int hieu=a-b;
   int tich=a*b;
-  int thuong=a/b;
+  float thuong=float(a/b);
   int du=a%b;
   cout << "Tong cua 2 so nguyen " << a << " va " << b << " la: " << tong << "\n";
   cout << "Hieu cua 2 so nguyen " << a << " va " << b << " la: " << hieu << "\n";
